+++ joyrpc-test/joyrpc-test-transport/src/test/java/io/joyrpc/transport/netty4/mock/ClientMockCodec.java
@@ -9,9 +9,9 @@ package io.joyrpc.transport.netty4.mock;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *     http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -58,7 +58,7 @@ public class ClientMockCodec implements LengthFieldFrameCodec {
         MockMessage mockMessage = (MockMessage) message;
         buffer.writeBytes(new byte[]{'M', 'O', 'C', 'K'});
         buffer.writeInt(mockMessage.getPayLoad().length + 13);
-        buffer.writeInt(mockMessage.getHeader().getMsgId());
+        buffer.writeInt((int) mockMessage.getHeader().getMsgId());
         buffer.writeByte(mockMessage.getMsgType());
         buffer.writeBytes(mockMessage.getPayLoad());
     }
